requirements.txt
fuse
python-magic
fusepy
gmusicapi
eyed3
Pillow==7.1.2
